refactor: rank 로직 리팩토링
